Put remote machines code in a separate file
